Initial commit. Project based in previous works
Project based in CppND-Route-Planning-Project from Udacity. Created to
practice basic use of cpp tools and train with A*search algorithm.
Improved readme to compile and make it works in windows and linux.

The installation of the IO2D libraries in windows is too complicated, so
to facilitate the compilation in windows it's recommended to use WLS2 or
linux virtual machines.